light simplification of terms
