+++ Project.toml
@@ -1,7 +1,7 @@
 name = "SimpleExpressions"
 uuid = "deba94f7-f32a-40ad-b45e-be020a5ded2f"
 authors = ["jverzani <jverzani@gmail.com> and contributors"]
-version = "1.1.9"
+version = "1.1.10"
 
 [deps]
 CallableExpressions = "391672e0-bbe4-4ab4-8bc9-b89a79cbc2f0"
@@ -22,6 +22,7 @@ SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 [extensions]
 SimpleExpressionsAbstractTreesExt = "AbstractTrees"
 SimpleExpressionsLatexifyExt = "Latexify"
+#SimpleExpressionsMetatheoryExt = "Metatheory"
 SimpleExpressionsRecipesBaseExt = "RecipesBase"
 SimpleExpressionsRootsExt = "Roots"
 SimpleExpressionsSpecialFunctionsExt = "SpecialFunctions"
