+++ src/scalar-derivative.jl
@@ -1,3 +1,13 @@
+## could name this diff(ex, x)
+## Maple       diff
+## Matlab      diff
+## Mathematica derivative
+## SymPy       diff
+## Sage        derivative
+## Polynomials derivative
+## Symbolics   Differential Differential(x) = Base.Fix2(SimpleExpressions.D,x)
+## Symbolics   derivative
+
 """
     D(::AbstractSymbolic, [x])
 
@@ -28,6 +38,8 @@ D(𝑥::SymbolicVariable, x) = 𝑥 == x ? 1 : 0
 D(𝑥::SymbolicParameter, x) = 𝑥 == x ? 1 : 0
 D(ex::SymbolicEquation, x) = D(ex.lhs, x) ~ D(ex.rhs, x)
 
+# combine slows this down
+#D(ex::SymbolicExpression, x) = combine(D(operation(ex), arguments(ex), x))
 D(ex::SymbolicExpression, x) = D(operation(ex), arguments(ex), x)
 
 
@@ -51,13 +63,13 @@ D(ex::SymbolicEquation) = D(ex.lhs) ~ D(ex.rhs)
 # cases
 ## sum rule
 function D(::typeof(+), args, x)
-    reduce(⊕, D.(args, x); init=zero(x))
+    reduce(+, D.(args, x); init=zero(x))
 end
 D(::typeof(sum), args, x) = SymbolicExpression(+, D.(args), x)
 
 ## difference rule
 function D(::typeof(-), args, x)
-    return reduce(⊖, D.(args, x); init=zero(x))
+    return reduce(-, D.(args, x); init=zero(x))
 end
 
 ## product rule
@@ -70,7 +82,7 @@ function D(::typeof(*), args, x)
             aa[i] = a
         end
         aa[i] = ai′
-        tot = tot ⊕ reduce(⊗, aa)
+        tot = tot + reduce(*, aa)
     end
     return tot
 end
@@ -80,7 +92,7 @@ D(::typeof(prod), args, x) = D(SymbolicExpression(*, args), x)
 function D(::typeof(/), args, x)
     u,v = args
     u′, v′ = D(u,x), D(v,x)
-    ((u′ ⊗ v) ⊖ (u ⊗ v′)) ⨸ (v⊗v)
+    ((u′ * v) - (u * v′)) / (v*v)
 end
 
 ## powers
@@ -89,9 +101,9 @@ function D(::typeof(^), args,x)
 
     if !contains(b, x)
         iszero(b) && return zero(x)
-        isone(b) && return D(a,x) ⊗ a
-        isone(b-1) && return D(a,x) ⊗ (2*a)
-        return  D(a,x) ⊗ (b*a^(b()-1))
+        isone(b) && return D(a,x) * a
+        isone(b-1) && return D(a,x) * (2*a)
+        return  D(a,x) * (b*a^(b()-1))
     else
         return D(exp(b * log(a)),x)
     end
@@ -103,46 +115,46 @@ end
 D(::typeof(sqrt), args,x) = (𝑥 = only(args); D(𝑥,x) / sqrt(𝑥) * (1//2))
 D(::typeof(cbrt), args,x) = (𝑥 = only(args); D(𝑥,x) / cbrt(𝑥)^2 * (1//3))
 
-D(::typeof(inv), args,x)     = (𝑥 = only(args); D(𝑥,x) ⊗ -1/𝑥^2 ⊗ 𝕀(Ne(𝑥,0)))
-D(::typeof(abs), args,x)     = (𝑥 = only(args); D(𝑥,x) ⊗ sign(𝑥) ⊗ 𝕀(Ne(𝑥, 0)))
-D(::typeof(sign), args,x)    = (𝑥 = only(args); 0 ⊗ 𝕀(𝑥 != 0))
-D(::typeof(abs2), args,x)    = (𝑥 = only(args); D(𝑥,x) ⊗ 2𝑥)
-D(::typeof(deg2rad), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ (pi / 180))
-D(::typeof(rad2deg), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ (180 / pi))
+D(::typeof(inv), args,x)     = (𝑥 = only(args); D(𝑥,x) * -1/𝑥^2 * 𝕀(Ne(𝑥,0)))
+D(::typeof(abs), args,x)     = (𝑥 = only(args); D(𝑥,x) * sign(𝑥) * 𝕀(Ne(𝑥, 0)))
+D(::typeof(sign), args,x)    = (𝑥 = only(args); 0 * 𝕀(𝑥 != 0))
+D(::typeof(abs2), args,x)    = (𝑥 = only(args); D(𝑥,x) * 2𝑥)
+D(::typeof(deg2rad), args,x) = (𝑥 = only(args); D(𝑥,x) * (pi / 180))
+D(::typeof(rad2deg), args,x) = (𝑥 = only(args); D(𝑥,x) * (180 / pi))
 
-D(::typeof(exp), args,x)   = (𝑥 = only(args); D(𝑥,x) ⊗ exp(𝑥))
-D(::typeof(exp2), args,x)  = (𝑥 = only(args); D(𝑥,x) ⊗ exp2(𝑥) ⊗ log(2))
-D(::typeof(exp10), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ exp10(𝑥) ⊗ log(10))
-D(::typeof(expm1), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ exp(𝑥))
-D(::typeof(log), args,x)   = (𝑥 = only(args); D(𝑥,x) ⊗ (1/𝑥) ⊗ 𝕀(Ge(𝑥,0)))
-D(::typeof(log2), args,x)  = (𝑥 = only(args); D(𝑥,x) ⊗ (1/𝑥/log(2)) ⊗ 𝕀(Ge(𝑥, 0)))
-D(::typeof(log10), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ (1/𝑥/log(10)) ⊗ 𝕀(Ge(𝑥, 0)))
-D(::typeof(log1p), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ 1/(1 + 𝑥))
+D(::typeof(exp), args,x)   = (𝑥 = only(args); D(𝑥,x) * exp(𝑥))
+D(::typeof(exp2), args,x)  = (𝑥 = only(args); D(𝑥,x) * exp2(𝑥) * log(2))
+D(::typeof(exp10), args,x) = (𝑥 = only(args); D(𝑥,x) * exp10(𝑥) * log(10))
+D(::typeof(expm1), args,x) = (𝑥 = only(args); D(𝑥,x) * exp(𝑥))
+D(::typeof(log), args,x)   = (𝑥 = only(args); D(𝑥,x) * (1/𝑥) * 𝕀(Ge(𝑥,0)))
+D(::typeof(log2), args,x)  = (𝑥 = only(args); D(𝑥,x) * (1/𝑥/log(2)) * 𝕀(Ge(𝑥, 0)))
+D(::typeof(log10), args,x) = (𝑥 = only(args); D(𝑥,x) * (1/𝑥/log(10)) * 𝕀(Ge(𝑥, 0)))
+D(::typeof(log1p), args,x) = (𝑥 = only(args); D(𝑥,x) * 1/(1 + 𝑥))
 
 
-D(::typeof(sin), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  cos(𝑥))
-D(::typeof(cos), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -sin(𝑥))
-D(::typeof(tan), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  sec(𝑥)^2)
-D(::typeof(sec), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  sec(𝑥) ⊗ tan(𝑥))
-D(::typeof(csc), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -csc(𝑥) ⊗ cot(𝑥))
-D(::typeof(cot), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -csc(𝑥)^2)
+D(::typeof(sin), args,x) = (𝑥 = only(args); D(𝑥,x) *  cos(𝑥))
+D(::typeof(cos), args,x) = (𝑥 = only(args); D(𝑥,x) * -sin(𝑥))
+D(::typeof(tan), args,x) = (𝑥 = only(args); D(𝑥,x) *  sec(𝑥)^2)
+D(::typeof(sec), args,x) = (𝑥 = only(args); D(𝑥,x) *  sec(𝑥) * tan(𝑥))
+D(::typeof(csc), args,x) = (𝑥 = only(args); D(𝑥,x) * -csc(𝑥) * cot(𝑥))
+D(::typeof(cot), args,x) = (𝑥 = only(args); D(𝑥,x) * -csc(𝑥)^2)
 
 D(::typeof(asin), args,x) = (𝑥 = only(args); D(𝑥,x) / sqrt(1 - 𝑥^2))
 D(::typeof(acos), args,x) = (𝑥 = only(args); D(𝑥,x) / (-sqrt(1 - 𝑥^2)))
 D(::typeof(atan), args,x) = (𝑥 = only(args); D(𝑥,x) / (1 + 𝑥^2))
-D(::typeof(asec), args,x) = (𝑥 = only(args); D(𝑥,x) / (abs(𝑥) ⊗ sqrt(𝑥^2 - 1)))
-D(::typeof(acsc), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ (abs(𝑥) ⊗ sqrt(𝑥^2 - 1)) ⊗ (-1))
-D(::typeof(acot), args,x) = (𝑥 = only(args); D(𝑥,x) / (1 + 𝑥^2) ⊗ (-1))
+D(::typeof(asec), args,x) = (𝑥 = only(args); D(𝑥,x) / (abs(𝑥) * sqrt(𝑥^2 - 1)))
+D(::typeof(acsc), args,x) = (𝑥 = only(args); D(𝑥,x) * (abs(𝑥) * sqrt(𝑥^2 - 1)) * (-1))
+D(::typeof(acot), args,x) = (𝑥 = only(args); D(𝑥,x) / (1 + 𝑥^2) * (-1))
 
-D(::typeof(sinh), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  cosh(𝑥))
-D(::typeof(cosh), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  sinh(𝑥))
-D(::typeof(tanh), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  sech(𝑥)^2)
-D(::typeof(sech), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -sech(𝑥) ⊗ tanh(𝑥))
-D(::typeof(csch), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -csch(𝑥) ⊗ coth(𝑥))
-D(::typeof(coth), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -csch(𝑥)^2)
+D(::typeof(sinh), args,x) = (𝑥 = only(args); D(𝑥,x) *  cosh(𝑥))
+D(::typeof(cosh), args,x) = (𝑥 = only(args); D(𝑥,x) *  sinh(𝑥))
+D(::typeof(tanh), args,x) = (𝑥 = only(args); D(𝑥,x) *  sech(𝑥)^2)
+D(::typeof(sech), args,x) = (𝑥 = only(args); D(𝑥,x) * -sech(𝑥) * tanh(𝑥))
+D(::typeof(csch), args,x) = (𝑥 = only(args); D(𝑥,x) * -csch(𝑥) * coth(𝑥))
+D(::typeof(coth), args,x) = (𝑥 = only(args); D(𝑥,x) * -csch(𝑥)^2)
 
-D(::typeof(sinpi), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗  π ⊗ cospi(𝑥))
-D(::typeof(cospi), args,x) = (𝑥 = only(args); D(𝑥,x) ⊗ -π ⊗ sinpi(𝑥))
+D(::typeof(sinpi), args,x) = (𝑥 = only(args); D(𝑥,x) *  π * cospi(𝑥))
+D(::typeof(cospi), args,x) = (𝑥 = only(args); D(𝑥,x) * -π * sinpi(𝑥))
 
 ## more in SpecialFunctions.jl extension
